--- a/src/bench/bls.cpp
+++ b/src/bench/bls.cpp
@@ -359,8 +359,8 @@ static void BLS_Verify_BatchedParallel(benchmark::State& state)
     }
 }
 
-BENCHMARK(BLS_PubKeyAggregate_Normal, 300 * 1000)
-BENCHMARK(BLS_SecKeyAggregate_Normal, 700 * 1000)
+BENCHMARK(BLS_PubKeyAggregate_Normal, 140 * 1000)
+BENCHMARK(BLS_SecKeyAggregate_Normal, 800 * 1000)
 BENCHMARK(BLS_SignatureAggregate_Normal, 100 * 1000)
 BENCHMARK(BLS_Sign_Normal, 600)
 BENCHMARK(BLS_Verify_Normal, 350)